Remove execution.rs
