+++ crates/sui-single-node-benchmark/src/benchmark_context.rs
@@ -169,6 +169,115 @@ impl BenchmarkContext {
         results.into_iter().map(|r| r.unwrap()).collect()
     }
 
+    pub(crate) async fn benchmark_transaction_execution(&self, transactions: Vec<Transaction>) {
+        let mut transactions = self.certify_transactions(transactions).await;
+        self.execute_sample_transaction(transactions.pop().unwrap())
+            .await;
+
+        let tx_count = transactions.len();
+        let start_time = std::time::Instant::now();
+        info!(
+            "Started executing {} transactions. You can now attach a profiler",
+            transactions.len()
+        );
+        self.execute_transactions(transactions).await;
+        let elapsed = start_time.elapsed().as_millis() as f64 / 1000f64;
+        info!(
+            "Execution finished in {}s, TPS={}",
+            elapsed,
+            tx_count as f64 / elapsed
+        );
+    }
+
+    /// Print out a sample transaction and its effects so that we can get a rough idea
+    /// what we are measuring.
+    async fn execute_sample_transaction(&self, sample_transaction: CertifiedTransaction) {
+        info!("Sample transaction: {:?}", sample_transaction.data());
+        let effects = self
+            .validator()
+            .execute_tx_immediately(sample_transaction.into_unsigned())
+            .await;
+        info!("Sample effects: {:?}\n\n", effects);
+        assert!(effects.status().is_ok());
+    }
+
+    /// Benchmark parallel signing a vector of transactions and measure the TPS.
+    pub(crate) async fn benchmark_transaction_signing(&self, transactions: Vec<Transaction>) {
+        let sample_transaction = &transactions[0];
+        info!("Sample transaction: {:?}", sample_transaction.data());
+
+        let tx_count = transactions.len();
+        let start_time = std::time::Instant::now();
+        self.validator_sign_transactions(transactions).await;
+        let elapsed = start_time.elapsed().as_millis() as f64 / 1000f64;
+        info!(
+            "Transaction signing finished in {}s, TPS={}.",
+            elapsed,
+            tx_count as f64 / elapsed,
+        );
+    }
+
+    pub(crate) async fn benchmark_checkpoint_executor(
+        &self,
+        transactions: Vec<Transaction>,
+        checkpoint_size: usize,
+    ) {
+        let mut transactions = self.certify_transactions(transactions).await;
+
+        self.execute_sample_transaction(transactions.pop().unwrap())
+            .await;
+
+        info!("Executing all transactions to generate effects");
+        let tx_count = transactions.len();
+        let effects: BTreeMap<_, _> = self
+            .execute_transactions(transactions.clone())
+            .await
+            .into_iter()
+            .map(|e| (*e.transaction_digest(), e))
+            .collect();
+        info!("Reverting all transactions so that we could re-execute through checkpoint executor");
+        self.revert_transactions(effects.keys()).await;
+
+        info!("Building checkpoints");
+        let validator = self.validator();
+        let checkpoints = validator
+            .build_checkpoints(transactions, effects, checkpoint_size)
+            .await;
+        info!("Built {} checkpoints", checkpoints.len());
+        let (mut checkpoint_executor, checkpoint_sender) = validator.create_checkpoint_executor();
+        for (checkpoint, contents) in checkpoints {
+            let state = validator.get_validator();
+            state
+                .get_checkpoint_store()
+                .insert_verified_checkpoint(&checkpoint)
+                .unwrap();
+            state
+                .database
+                .multi_insert_transaction_and_effects(contents.iter())
+                .unwrap();
+            state
+                .get_checkpoint_store()
+                .insert_verified_checkpoint_contents(&checkpoint, contents)
+                .unwrap();
+            state
+                .get_checkpoint_store()
+                .update_highest_synced_checkpoint(&checkpoint)
+                .unwrap();
+            checkpoint_sender.send(checkpoint).unwrap();
+        }
+        let start_time = std::time::Instant::now();
+        info!("Starting checkpoint execution. You can now attach a profiler");
+        checkpoint_executor
+            .run_epoch(validator.get_epoch_store().clone())
+            .await;
+        let elapsed = start_time.elapsed().as_millis() as f64 / 1000f64;
+        info!(
+            "Checkpoint execution finished in {}s, TPS={}.",
+            elapsed,
+            tx_count as f64 / elapsed,
+        );
+    }
+
     pub(crate) async fn execute_transactions(
         &self,
         transactions: Vec<CertifiedTransaction>,
